+++ lib/absinthe/language/fragment.ex
@@ -36,11 +36,4 @@ defmodule Absinthe.Language.Fragment do
       Blueprint.SourceLocation.at(loc)
     end
   end
-
-  defimpl Absinthe.Traversal.Node do
-    def children(node, _schema) do
-      [node.directives, List.wrap(node.selection_set)]
-      |> Enum.concat()
-    end
-  end
 end